Added BufferStream implementation

+++ core/src/cubos/memory/std_stream.cpp
@@ -29,7 +29,7 @@ size_t StdStream::tell() const
     return ftell(this->file);
 }
 
-void StdStream::seek(size_t offset, SeekOrigin origin)
+void StdStream::seek(int64_t offset, SeekOrigin origin)
 {
     switch (origin)
     {
